Decommonize transferMemory from native target
